+++ opsml/registry/sql/base/registry_base.py
@@ -2,7 +2,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 import uuid
-from typing import Any, Dict, List, Optional, Tuple, cast, Sequence
+from typing import Any, Dict, List, Optional, Sequence, Tuple, cast
 
 from semver import VersionInfo
